Move fetching of files and parameters to aws package
